--- src/hooks/useSubcategories.ts
@@ -1,17 +1,0 @@
-import { useEffect } from "react";
-import { useAppDispatch } from "../redux/hooks/useAppDispatch";
-import { useAppSelector } from "../redux/hooks/useAppSelector";
-import { fetchSubcategoriesThunk } from "../redux/subcategories/subcategoriesOperations";
-import { selectSubcategories } from "../redux/subcategories/subcategoriesSlice";
-
-export default function useSubcategories() {
-  const dispatch = useAppDispatch();
-
-  useEffect(() => {
-    dispatch(fetchSubcategoriesThunk());
-  }, [dispatch]);
-
-  const subcategories = useAppSelector(selectSubcategories);
-
-  return subcategories;
-}
